refactor: 회원가입 로직 리팩터링 및 응답 구조 개선

See merge request s12-bigdata-recom-sub1/S12P21A304!75

+++ backend/src/main/java/newstock/domain/user/repository/UserCustomRepository.java
@@ -6,5 +6,5 @@ import java.util.Optional;
 
 public interface UserCustomRepository {
 
-    Optional<User> findById(Long userId);
+    Optional<User> findById(Integer userId);
 }
